moneymong-172 feat: ocr captureview fix
